philo_one working using multiple solution and size_t on ms! needed norminette

+++ philo_one_remastered/includes/utils.h
@@ -6,7 +6,7 @@
 /*   By: lmartin <lmartin@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/06/09 00:07:39 by lmartin           #+#    #+#             */
-/*   Updated: 2020/06/14 01:10:58 by lmartin          ###   ########.fr       */
+/*   Updated: 2020/06/16 02:15:37 by lmartin          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -19,7 +19,7 @@ typedef struct		s_parameters
 	size_t			time_to_die;
 	size_t			time_to_eat;
 	size_t			time_to_sleep;
-	int				number_of_time_each_philosophers_must_eat;
+	size_t			number_of_time_each_philosophers_must_eat;
 	struct timeval	*time_start;
 }					t_parameters;
